# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenISMS API",
    "description": "REST API für das Information Security Management System (OpenISMS).\n\n## Authentifizierung\nAlle geschützten Endpunkte erfordern einen **Bearer JWT-Token** im `Authorization`-Header:\n```\nAuthorization: Bearer <token>\n```\nDen Token erhalten Sie über `POST /api/auth/login`.\n\n## Rollen\n- `admin` – Vollzugriff\n- `assessor` – Risikobewerter / Auditor\n- `it-staff` – IT-Mitarbeiter (Security-Daten)\n- `dpo` – Datenschutzbeauftragter\n- `owner` – Asset-Verantwortlicher\n- `management` – Nur-Lesen (Management-Ebene)\n- `employee` – Schulungs-Tab, eigene Trainingsübersicht\n- `viewer` – Nur Lesezugriff\n\n**Custom Roles**: Eigene Rollennamen mit Basisrolle über `/api/admin/custom-roles` verwalten.",
    "version": "2.2.0",
    "contact": { "name": "ISMS Support" }
  },
  "servers": [
    { "url": "/api", "description": "Aktuelle Instanz" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin","assessor","it-staff","dpo","owner","management","employee","viewer"] },
          "department": { "type": "string" },
          "active": { "type": "boolean" },
          "sso_user": { "type": "boolean" },
          "totp_enabled": { "type": "boolean" },
          "custom_role_id": { "type": "integer", "nullable": true },
          "customRole": { "$ref": "#/components/schemas/CustomRole" },
          "failed_login_attempts": { "type": "integer" },
          "lockout_until": { "type": "string", "format": "date-time", "nullable": true },
          "last_seen_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "CustomRole": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "base_role": { "type": "string", "enum": ["admin","assessor","it-staff","dpo","owner","management","employee","viewer"] },
          "users_count": { "type": "integer" }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "color": { "type": "string", "example": "#3b82f6" },
          "created_by_id": { "type": "integer", "nullable": true },
          "members": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/assets/{id}/documents": {
      "get": {
        "tags": ["Assets"],
        "summary": "Dokumente eines Assets abrufen",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Dokumentenliste" } }
      },
      "post": {
        "tags": ["Assets"],
        "summary": "Dokument hochladen (max. 25 MB)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" }, "category": { "type": "string" }, "description": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Dokument gespeichert" } }
      }
    },
...
    "/groups": {
      "get": {
        "tags": ["Gruppen"],
        "summary": "Alle Gruppen abrufen",
        "responses": { "200": { "description": "Gruppenliste mit Mitgliedern", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Group" } } } } } }
      },
      "post": {
        "tags": ["Gruppen"],
        "summary": "Neue Gruppe anlegen",
        "description": "Erfordert Rolle: admin",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "color": { "type": "string", "example": "#3b82f6" } } } } }
        },
        "responses": { "201": { "description": "Gruppe erstellt" } }
      }
    },
    "/groups/{id}": {
      "get": {
        "tags": ["Gruppen"],
        "summary": "Gruppe mit Mitgliedern abrufen",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Gruppe" } }
      },
      "put": {
        "tags": ["Gruppen"],
        "summary": "Gruppe aktualisieren",
        "description": "Erfordert Rolle: admin",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Group" } } } },
        "responses": { "200": { "description": "Gruppe aktualisiert" } }
      },
      "delete": {
        "tags": ["Gruppen"],
        "summary": "Gruppe löschen",
        "description": "Erfordert Rolle: admin",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Gruppe gelöscht" } }
      }
    },
    "/groups/{id}/members": {
      "post": {
        "tags": ["Gruppen"],
        "summary": "Mitglied zur Gruppe hinzufügen",
        "description": "Erfordert Rolle: admin",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["user_id"], "properties": { "user_id": { "type": "integer" } } } } } },
        "responses": { "200": { "description": "Mitglied hinzugefügt" } }
      }
    },
    "/groups/{id}/members/{userId}": {
      "delete": {
        "tags": ["Gruppen"],
        "summary": "Mitglied aus Gruppe entfernen",
        "description": "Erfordert Rolle: admin",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Mitglied entfernt" } }
      }
    },
...
  },
...
}